# CookFlowVue/openapi.yaml
openapi: 3.0.3
info:
  title: CookFlow2 API
  version: 1.0.0
  description: API spec generated from frontend for CookFlow2 (for generating Spring Boot backend)
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
...
  /users/{username}:
    get:
      summary: Get user profile
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
      responses:
        '200':
          description: user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
    put:
      summary: Update user profile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfile'
      responses:
        '200':
          description: updated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserProfile:
      type: object
      properties:
        username:
          type: string
        avatar:
          type: string
        bio:
          type: string
        favorites:
          type: array
          items:
            type: integer
...